fix: add try-catch around user fetch when exporting poll results to avoid unknown member error
Exporting poll reasons would give an error "Unknown Member", with logs indicating there is a 10007 error from Discord. This only affects older polls and not newer polls in Nouncil. I suspect the reason is that a user left the guild and we are unable to retrieve the username.

This change lets us still export polls when that happens, by displaying the user's Discord ID instead (graceful failure).